pyproject.toml
[project]
name = "ai-loop"
version = "0.1.0"
description = "Simple resilient loop runner for AI terminal tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
ailoop = "ailoop.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ailoop"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
